sdhgf

+++ Arkanoid/src/app/BonusPlayer.java
@@ -40,7 +40,7 @@ public class BonusPlayer extends Bonus {
     }
 
     @Override
-    public int aceleracion() {
+    public double aceleracion() {
         return 0;
     }
