+++ Arkanoid/src/app/BonusSlow.java
@@ -41,7 +41,7 @@ public class BonusSlow extends Bonus {
     }
 
     @Override
-    public int aceleracion() {
+    public double aceleracion() {
         return 0;
     }
