+++ Arkanoid/src/app/BonusWarp.java
@@ -42,7 +42,7 @@ public class BonusWarp extends Bonus {
     }
 
     @Override
-    public int aceleracion() {
+    public double aceleracion() {
         return 0;
     }
