+++ Arkanoid/src/app/Movible.java
@@ -3,5 +3,5 @@ package app;
 public interface Movible{
     public void mover() ;  //mover
     public double velocidad() ;//velocidad
-    public int aceleracion() ;//aceleracion
+    public double aceleracion() ;//aceleracion
 }
